% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntBias.R
\name{EntBias}
\alias{EntBias}
\title{Bias of entropy estimator}
\usage{
EntBias(X, k, sd)
}
\arguments{
\item{X}{the sample}

\item{k}{the order of nearest neighbour distances to work out}

\item{sd}{the standard deviation of the normal distribution the sample is from}
}
\description{
Works out the bias of the Kozachenko Leonenko estimator for entropy against
the actual value, given that the sample is from a normal distribution
}

