% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpoEnt.R
\name{ExpoEnt}
\alias{ExpoEnt}
\title{The actual value of entropy for the exponential distribution}
\usage{
ExpoEnt()
}
\arguments{
\item{rate}{the rate parameter of the exponential distribution}
}
\description{
The actual value of entropy for the exponential distribution
}

