% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLEE.R
\name{KLEE}
\alias{KLEE}
\title{KLEE1 - Kozachenko-Leonenko Estimator of Entropy}
\usage{
KLEE(X, k, d = 1)
}
\arguments{
\item{X}{a vector of samples from a known/unknown distribution}

\item{k}{the degree of nearest neigbour to estimate by}

\item{d}{the dimension of the sample}
}
\description{
Uses the kth Nearest Neighbour method, expanded by Kozachenko and Leonenko to generate a constsitent and asymptotically unbiased estimator for entropy
Only considers the 1 dimensional case (so far)
}

