% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalEntropy.R
\name{NormalEnt}
\alias{NormalEnt}
\title{The actual value of entropy for the normal distribution}
\usage{
NormalEnt(sd)
}
\arguments{
\item{sd}{the standard deviation of the normal distribution}
}
\description{
The actual value of entropy for the normal distribution
}

