% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rho.R
\name{Rho}
\alias{Rho}
\title{Distance of kth nearest neighbour}
\usage{
Rho(X, k, d = 1)
}
\arguments{
\item{X}{the sample to work out the distance between a value and its kth 
nearest neighbour}

\item{k}{the order of nearest neighbour to be used}

\item{d}{the dimension of the sample}
}
\description{
Distance of kth nearest neighbour
}

