% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roe.R
\name{Roe}
\alias{Roe}
\title{Distance of kth nearest neighbour}
\usage{
Roe(X, k, d = 1)
}
\arguments{
\item{X}{the sample from a distribution}

\item{k}{the order of nearest neighbour to be used}

\item{d}{the dimension of the sample}
}
\description{
Distance of kth nearest neighbour
}

