% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplesBias.R
\name{SamplesBias}
\alias{SamplesBias}
\title{A function to return the average value of N KLEE estimators}
\usage{
SamplesBias(N = 5000, dist = c("normal", "uniform", "exponential"), k = 1,
  M = 500, sd = 1, min = 0, max = 1, rate = 1)
}
\arguments{
\item{N}{sample size}

\item{dist, }{distribution for sample to be taken from, either "normal"
or "uniform"}

\item{M}{number of estimators}

\item{sd, }{standard deviation for sample from normal distubution, only
use if dist = "normal"}

\item{min, }{minimum for uniform distribution}

\item{max, }{maximum for uniform distribution}
}
\description{
A function to return the average value of N KLEE estimators
}

