% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniformEntropy.R
\name{UniformEnt}
\alias{UniformEnt}
\title{The actual value of entropy for the unfiorm distribution}
\usage{
UniformEnt(min, max)
}
\arguments{
\item{min}{the lowest number to take the distribution from}

\item{max}{the highest number to take the distribution to}
}
\description{
The actual value of entropy for the unfiorm distribution
}

