% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasRegression.R
\name{BiasRegression}
\alias{BiasRegression}
\title{A function to plot simulations for the Kozachenko-Leonenko entropy estimator}
\usage{
BiasRegression(k = 1, N = 5000, sd = 1)
}
\arguments{
\item{k}{the number for nearest neighbour, default 1}

\item{N}{the size of the largest sample to be considered, default 50000}

\item{sd}{the standard deviation for the normal distribution that the 
sample is from, default 1}
}
\description{
This function considers the logarithm of the Bias of entropy estimator for a 
sample from the normal distribution, against the logarithm of the size of 
this sample. It also plots a regression line and returns information about
the linear regression equation; Bias(H) = c/(N^a)
}

