% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasRegression.R
\name{BiasRegression}
\alias{BiasRegression}
\title{A function to plot simulations for the Kozachenko-Leonenko entropy estimator}
\usage{
BiasRegression(dist = c("normal", "uniform", "exponential"), k = 1,
  N = 25000, sd = 1, min = 0, max = 1, rate = 1, M = 500)
}
\arguments{
\item{dist, }{the distribution for the simulation to come from, either 
normal or uniform}

\item{k}{the number for nearest neighbour, default 1}

\item{N}{the size of the largest sample to be considered, default 250000}

\item{sd}{the standard deviation for the normal distribution that the 
sample is from, given the distribution is "normal", default 1}

\item{max/min, }{the parameters for the unifrom distribution, provided
that the distribution is "uniform", default is min=0, max=1}
}
\description{
This function considers the logarithm of the Bias of entropy estimator for a 
sample from the normal or uniform distribution, against the logarithm of 
the size of this sample. It also plots a regression line and returns 
information about the linear regression equation; Bias(H) = c/(N^a)
}

